requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
eva-decord==0.6.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kaggle==1.7.4.5
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==22.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
torchviz==0.0.3
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wget==3.2
xxhash==3.6.0
yarl==1.22.0
